Extract sync/async flags/enums to aioenums
